refactor: split useMemoryBooks into 3 sub-composables (T8)
useMemoryBooks (806 → ~230 lines orchestrator) decomposed into:
- useMemoryDraftProgress: draft generation lifecycle, abort controllers,
  timer-based progress tracking (~115 lines)
- useMemoryIndexing: search type picker, vector toggle, reindex all
  (~160 lines)
- useMemoryCRUD: approve/delete drafts, delete entries, cancel draft,
  maintenance sheet (~210 lines)

useMemoryBooks now composes from sub-composables, passes shared deps
(currentMemoryBookData ref, loadCurrentMemoryBook, etc.) via DI.
handleMemoryScanChat remains in the orchestrator due to tight coupling
with both draft state and patchChatData.

Public API unchanged — ChatView.vue imports unchanged.
